+++ cell.h
@@ -11,10 +11,10 @@ class cell {
 public:
     cell();
     //~cell(void);
-    void setData(int popSick, int inf, int imf, int tIn, int tIm);
+    void setData(double popSick, int inf, int imf, int tIn, int tIm);
     void setPostion(int i, int j);
 
-    int popSick; //Population sickness =  infected population/ total population
+    double popSick; //Population sickness =  infected population/ total population
     int inf;    //infection flag
     int imf;    //immune flag
     int tIn;    //time infected
